@@ -6,7 +6,6 @@ import { Text, Header, Body, Title } from 'native-base';
 import {
 	ReactiveBase,
 	DataSearch,
-	TextField,
 	SingleDropdownList,
 	MultiDropdownList,
 	SingleDropdownRange,
@@ -21,7 +20,7 @@ import {
 class Main extends Component {
 	state = {
 		isReady: false,
-	}
+	};
 
 	async componentWillMount() {
 		await Expo.Font.loadAsync({
@@ -40,12 +39,12 @@ class Main extends Component {
 			keyExtractor={item => item._id}
 			renderItem={({ item }) => (
 				<View style={{ margin: 5 }}>
-					<Text
-						style={{ flex: 1, fontWeight: 'bold' }}
-					>
+					<Text style={{ flex: 1, fontWeight: 'bold' }}>
 						{this.parseToElement(item.name)}
 					</Text>
-					<Text>{item.brand} - {item.model}</Text>
+					<Text>
+						{item.brand} - {item.model}
+					</Text>
 				</View>
 			)}
 			onEndReachedThreshold={0.5}
@@ -72,7 +71,7 @@ class Main extends Component {
 		}
 
 		return str;
-	}
+	};
 
 	render() {
 		if (!this.state.isReady) {
@@ -81,9 +80,8 @@ class Main extends Component {
 
 		return (
 			<ReactiveBase
-				app="car-store"
-				credentials="cf7QByt5e:d2d60548-82a9-43cc-8b40-93cbbe75c34c"
-				type="cars"
+				app="carstore-dataset"
+				credentials="4HWI27QmA:58c731f7-79ab-4f55-a590-7e15c7e36721"
 				// theme={{
 				// 	textColor: 'yellow',
 				// }}
@@ -96,40 +94,18 @@ class Main extends Component {
 				<ScrollView>
 					<View style={{ padding: 10 }}>
 						<SelectedFilters />
-						<SingleDropdownList
-							componentId="StComponent"
-							dataField="brand.raw"
-						/>
+						<SingleDropdownList componentId="StComponent" dataField="brand.keyword" />
 
 						<MultiDropdownList
 							componentId="MultiDropdownListComponent"
-							dataField="brand.raw"
+							dataField="brand.keyword"
 							selectAllLabel="All"
 						/>
 
-						<TextField
-							componentId="TextFieldComponent"
-							dataField="color"
-							placeholder="Search color"
-							innerStyle={{
-								icon: {
-									color: 'blue',
-								},
-							}}
-							innerProps={{
-								icon: {
-									color: 'yellow',
-								},
-							}}
-						/>
-
-						<DatePicker
-							dataField="someField"
-							componentId="DatePicker"
-						/>
+						<DatePicker dataField="year" componentId="DatePicker" />
 
 						<DateRange
-							dataField="someField"
+							dataField="year"
 							componentId="DateRange"
 							defaultSelected={{
 								start: '2017-04-04',
@@ -177,7 +153,7 @@ class Main extends Component {
 
 						<DataSearch
 							componentId="DataSearchComponent"
-							dataField="name"
+							dataField="model"
 							defaultSelected="Nissan"
 							react={{
 								and: 'TextFieldComponent',
@@ -185,17 +161,14 @@ class Main extends Component {
 						/>
 
 						<ReactiveList
-							dataField="name"
+							dataField="model"
 							componentId="ReactiveList"
 							size={20}
 							onAllData={this.onAllData}
 							pagination
 							stream
 							react={{
-								and: [
-									'DataSearchComponent',
-									'StComponent',
-								],
+								and: ['DataSearchComponent', 'StComponent'],
 							}}
 						/>
 					</View>
