@@ -13,15 +13,15 @@ class Main extends Component {
 	render() {
 		return (
 			<ReactiveBase
-				app="car-store"
-				credentials="cf7QByt5e:d2d60548-82a9-43cc-8b40-93cbbe75c34c"
+				app="good-books-ds"
+				credentials="nY6NNTZZ6:27b76b9f-18ea-456c-bc5e-3a5263ebc63d"
 			>
 				<div className="row">
 					<div className="col">
 						<SelectedFilters />
 						<RatingsFilter
 							componentId="RatingsSensor"
-							dataField="rating"
+							dataField="average_rating_rounded"
 							title="RatingsFilter"
 							data={[
 								{ start: 4, end: 5, label: '4 stars and up' },
@@ -39,7 +39,7 @@ class Main extends Component {
 							title="Results"
 							from={0}
 							size={20}
-							renderData={this.renderData}
+							renderData={this.onData}
 							react={{
 								and: 'RatingsSensor',
 							}}
@@ -50,15 +50,37 @@ class Main extends Component {
 		);
 	}
 
-	renderData(res) {
-		const result = {
-			image:
-				'https://www.enterprise.com/content/dam/global-vehicle-images/cars/FORD_FOCU_2012-1.png',
-			title: res.name,
-			rating: res.rating,
-			description: res.brand,
+	onData(data) {
+		return {
+			title: (
+				<div
+					className="book-title"
+					dangerouslySetInnerHTML={{ __html: data.original_title }}
+				/>
+			),
+			description: (
+				<div className="flex column justify-space-between">
+					<div>
+						<div>
+							by <span className="authors-list">{data.authors}</span>
+						</div>
+						<div className="ratings-list flex align-center">
+							<span className="stars">
+								{Array(data.average_rating_rounded)
+									.fill('x')
+									.map((item, index) => (
+										<i className="fas fa-star" key={`star-${index + 1}`} />
+									)) // eslint-disable-line
+								}
+							</span>
+							<span className="avg-rating">({data.average_rating} avg)</span>
+						</div>
+					</div>
+					<span className="pub-year">Pub {data.original_publication_year}</span>
+				</div>
+			),
+			image: data.image,
 		};
-		return result;
 	}
 }
 
