@@ -13,9 +13,8 @@ import Layout from '../components/Layout';
 import ListItemView from '../components/ListItemView';
 
 const settings = {
-	app: 'meetup_demo',
-	credentials: 'LPpISlEBe:2a8935f5-0f63-4084-bc3e-2b2b4d1a8e02',
-	type: 'meetupdata1',
+	app: 'meetup_app',
+	credentials: 'lW70IgSjr:87c5ae16-73fb-4559-a29e-0a02760d2181',
 };
 
 const singleDataListProps = {
