@@ -6,11 +6,14 @@ import {
 	watchComponent,
 	updateQuery,
 	setQueryListener,
+	setQueryOptions,
 } from '@appbaseio/reactivecore/lib/actions';
 import {
 	checkValueChange,
 	checkPropChange,
 	getClassName,
+	pushToAndClause,
+	getQueryOptions,
 } from '@appbaseio/reactivecore/lib/utils/helper';
 
 import types from '@appbaseio/reactivecore/lib/utils/types';
@@ -20,6 +23,7 @@ import Input from '../../styles/Input';
 import Container from '../../styles/Container';
 import { UL, Radio } from '../../styles/FormControlList';
 import { connect } from '../../utils';
+import { getAggsQuery } from './utils';
 
 class SingleDataList extends Component {
 	constructor(props) {
@@ -29,14 +33,24 @@ class SingleDataList extends Component {
 		this.state = {
 			currentValue,
 			searchTerm: '',
+			options: props.data || [],
 		};
+
+		this.internalComponent = `${props.componentId}__internal`;
 		this.locked = false;
 
 		props.addComponent(props.componentId);
+		props.addComponent(this.internalComponent);
 		props.setQueryListener(props.componentId, props.onQueryChange, null);
 
 		this.setReact(props);
 		const hasMounted = false;
+
+		if (props.showCount) {
+			this.updateQueryOptions(props);
+		}
+
+		this.setReact(props);
 
 		if (currentValue) {
 			this.setValue(currentValue, props, hasMounted);
@@ -48,6 +62,20 @@ class SingleDataList extends Component {
 
 		checkPropChange(this.props.dataField, prevProps.dataField, () => {
 			this.updateQuery(this.state.currentValue, this.props);
+
+			if (this.props.showCount) {
+				this.updateQueryOptions(this.props);
+			}
+		});
+
+		checkPropChange(this.props.data, prevProps.data, () => {
+			if (this.props.showCount) {
+				this.updateQueryOptions(this.props);
+			}
+		});
+
+		checkPropChange(this.props.options, prevProps.options, () => {
+			this.updateStateOptions(this.props.options[this.props.dataField].buckets);
 		});
 
 		if (this.props.value !== prevProps.value) {
@@ -62,11 +90,18 @@ class SingleDataList extends Component {
 
 	componentWillUnmount() {
 		this.props.removeComponent(this.props.componentId);
+		this.props.removeComponent(this.internalComponent);
 	}
 
 	setReact(props) {
-		if (props.react) {
-			props.watchComponent(props.componentId, props.react);
+		const { react } = props;
+		if (react) {
+			const newReact = pushToAndClause(react, this.internalComponent);
+			props.watchComponent(props.componentId, newReact);
+		} else {
+			props.watchComponent(props.componentId, {
+				and: this.internalComponent,
+			});
 		}
 	}
 
@@ -141,6 +176,45 @@ class SingleDataList extends Component {
 		});
 	};
 
+	static generateQueryOptions(props, state) {
+		const queryOptions = getQueryOptions(props);
+		const includes = state.options.map(item => item.value);
+		return getAggsQuery(queryOptions, props, includes);
+	}
+
+	updateQueryOptions = (props) => {
+		const queryOptions = SingleDataList.generateQueryOptions(props, this.state);
+		props.setQueryOptions(this.internalComponent, queryOptions);
+	};
+
+	updateStateOptions = (bucket) => {
+		if (bucket) {
+			const bucketDictionary = bucket.reduce(
+				(obj, item) => ({
+					...obj,
+					[item.key]: item.doc_count,
+				}),
+				{},
+			);
+
+			const { options } = this.state;
+			const newOptions = options.map((item) => {
+				if (bucketDictionary[item.value]) {
+					return {
+						...item,
+						count: bucketDictionary[item.value],
+					};
+				}
+
+				return item;
+			});
+
+			this.setState({
+				options: newOptions,
+			});
+		}
+	};
+
 	handleInputChange = (e) => {
 		const { value } = e.target;
 		this.setState({
@@ -176,9 +250,10 @@ class SingleDataList extends Component {
 	};
 
 	render() {
-		const { selectAllLabel } = this.props;
+		const { selectAllLabel, showCount, renderListItem } = this.props;
+		const { options } = this.state;
 
-		if (this.props.data.length === 0) {
+		if (options.length === 0) {
 			return null;
 		}
 
@@ -215,7 +290,7 @@ class SingleDataList extends Component {
 							</label>
 						</li>
 					)}
-					{this.props.data
+					{options
 						.filter((item) => {
 							if (this.props.showSearch && this.state.searchTerm) {
 								return item.label
@@ -245,7 +320,25 @@ class SingleDataList extends Component {
 									className={getClassName(this.props.innerClass, 'label') || null}
 									htmlFor={`${this.props.componentId}-${item.label}`}
 								>
-									{item.label}
+									{renderListItem ? (
+										renderListItem(item.label, item.count)
+									) : (
+										<span>
+											{item.label}
+											{showCount && item.count && (
+												<span
+													className={
+														getClassName(
+															this.props.innerClass,
+															'count',
+														) || null
+													}
+												>
+													&nbsp;({item.count})
+												</span>
+											)}
+										</span>
+									)}
 								</label>
 							</li>
 						))}
@@ -259,9 +352,11 @@ SingleDataList.propTypes = {
 	addComponent: types.funcRequired,
 	removeComponent: types.funcRequired,
 	setQueryListener: types.funcRequired,
+	setQueryOptions: types.funcRequired,
 	updateQuery: types.funcRequired,
 	watchComponent: types.funcRequired,
 	selectedValue: types.selectedValue,
+	options: types.options,
 	// component props
 	beforeValueChange: types.func,
 	className: types.string,
@@ -286,6 +381,8 @@ SingleDataList.propTypes = {
 	themePreset: types.themePreset,
 	title: types.title,
 	URLParams: types.bool,
+	showCount: types.bool,
+	renderListItem: types.func,
 };
 
 SingleDataList.defaultProps = {
@@ -296,6 +393,7 @@ SingleDataList.defaultProps = {
 	showSearch: true,
 	style: {},
 	URLParams: false,
+	showCount: false,
 };
 
 const mapStateToProps = (state, props) => ({
@@ -304,6 +402,7 @@ const mapStateToProps = (state, props) => ({
 			&& state.selectedValues[props.componentId].value)
 		|| null,
 	themePreset: state.config.themePreset,
+	options: state.aggregations[props.componentId],
 });
 
 const mapDispatchtoProps = dispatch => ({
@@ -313,6 +412,7 @@ const mapDispatchtoProps = dispatch => ({
 	watchComponent: (component, react) => dispatch(watchComponent(component, react)),
 	setQueryListener: (component, onQueryChange, beforeQueryChange) =>
 		dispatch(setQueryListener(component, onQueryChange, beforeQueryChange)),
+	setQueryOptions: (component, props) => dispatch(setQueryOptions(component, props)),
 });
 
 export default connect(
