@@ -23,29 +23,33 @@ class ResultCard extends Component {
 					onClick={triggerClickAnalytics}
 					{...result.containerProps}
 				>
-					<Image
-						style={{ backgroundImage: `url(${result.image})` }}
-						className={getClassName(this.props.innerClass, 'image')}
-					/>
-					{typeof result.title === 'string' ? (
-						<Title
-							dangerouslySetInnerHTML={{ __html: result.title }}
-							className={getClassName(this.props.innerClass, 'title')}
+					{result.image && (
+						<Image
+							style={{ backgroundImage: `url(${result.image})` }}
+							className={getClassName(this.props.innerClass, 'image')}
 						/>
-					) : (
-						<Title className={getClassName(this.props.innerClass, 'title')}>
-							{result.title}
-						</Title>
 					)}
-					{typeof result.description === 'string' ? (
-						<article
-							dangerouslySetInnerHTML={{
-								__html: result.description,
-							}}
-						/>
-					) : (
-						<article>{result.description}</article>
-					)}
+					{result.title
+						&& (typeof result.title === 'string' ? (
+							<Title
+								dangerouslySetInnerHTML={{ __html: result.title }}
+								className={getClassName(this.props.innerClass, 'title')}
+							/>
+						) : (
+							<Title className={getClassName(this.props.innerClass, 'title')}>
+								{result.title}
+							</Title>
+						))}
+					{result.description
+						&& (typeof result.description === 'string' ? (
+							<article
+								dangerouslySetInnerHTML={{
+									__html: result.description,
+								}}
+							/>
+						) : (
+							<article>{result.description}</article>
+						))}
 				</Card>
 			);
 		}
