@@ -12,17 +12,14 @@ const onResultStats = (results, time) => (
 );
 
 const onData = data => ({
-	image:
-		data.vehicleType === 'other' || data.vehicleType === 'unknown'
-			? 'src/images/car.jpg'
-			: `src/images/${data.vehicleType.replace(/ /g, '-')}/${data.color}.jpg`,
-	title: data.name,
+	image: data.image,
+	title: data.model,
 	description: (
 		<div>
 			<div className={price}>${data.price}</div>
 			<Flex justifyContent="space-between" responsive>
 				<FlexChild>{'⭐'.repeat(data.rating)}</FlexChild>
-				<FlexChild>REGD. {data.yearOfRegistration}</FlexChild>
+				<FlexChild>REGD. {data.year}</FlexChild>
 			</Flex>
 			<Flex style={{ marginTop: 5 }} flexWrap>
 				{data.fuelType && <Topic>{data.fuelType}</Topic>}
@@ -36,8 +33,8 @@ const onData = data => ({
 const Results = () => (
 	<ResultCard
 		componentId="results"
-		dataField="name"
 		renderData={onData}
+		dataField="model"
 		onResultStats={onResultStats}
 		react={{
 			and: ['category', 'brand', 'rating', 'vehicle', 'price'],
