@@ -7,17 +7,19 @@ import { dataSearchContainer } from '../styles/Container';
 
 const Header = () => (
 	<Navbar>
-		<Flex
-			alignCenter
-			responsive
-			justifyContent="space-between"
-			className={navbarContent}
-		>
+		<Flex alignCenter responsive justifyContent="space-between" className={navbarContent}>
 			<FlexChild className={title}>Product Search</FlexChild>
 			<FlexChild className={dataSearchContainer}>
 				<DataSearch
 					componentId="search"
-					dataField={['name', 'tagline']}
+					dataField={[
+						'name',
+						'tagline',
+						'name.autosuggest',
+						'name.keyword',
+						'tagline.autosuggest',
+						'tagline.keyword',
+					]}
 					iconPosition="left"
 					placeholder="Discover products..."
 					URLParams
